+++ includes/Loop/Pagination.php
@@ -5,12 +5,14 @@
  * @package    Zooey
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.7
  */
 
 namespace WebManDesign\Zooey\Loop;
 
 use WebManDesign\Zooey\Component_Interface;
+use WebManDesign\Zooey\Tool\Arrow;
 use WP_HTML_Tag_Processor;
 
 // Exit if accessed directly.
@@ -21,7 +23,8 @@ class Pagination implements Component_Interface {
 	/**
 	 * Initialization.
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.0.7
 	 *
 	 * @return  void
 	 */
@@ -32,7 +35,7 @@ class Pagination implements Component_Interface {
 			// Filters
 
 				add_filter( 'paginate_links_output', __CLASS__ . '::paginate_links_output' );
-				add_filter( 'paginate_links_output', __CLASS__ . '::replace_arrows' );
+				add_filter( 'paginate_links_output', 'WebManDesign\Zooey\Tool\Arrow::replace' );
 
 				add_filter( 'render_block', __CLASS__ . '::render__pagination', ZOOEY_RENDER_BLOCK_PRIORITY, 2 );
 
@@ -60,51 +63,6 @@ class Pagination implements Component_Interface {
 	} // /paginate_links_output
 
 	/**
-	 * Replacing text arrows and chevron with SVG icons.
-	 *
-	 * @since  1.0.0
-	 *
-	 * @param  string $html
-	 *
-	 * @return  string
-	 */
-	public static function replace_arrows( string $html ): string {
-
-		// Variables
-
-			$svg = array(
-
-				'→' => '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="m224.49 136.49l-72 72a12 12 0 0 1-17-17L187 140H40a12 12 0 0 1 0-24h147l-51.49-51.52a12 12 0 0 1 17-17l72 72a12 12 0 0 1-.02 17.01"/></svg>',
-
-				'←' => '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M228 128a12 12 0 0 1-12 12H69l51.52 51.51a12 12 0 0 1-17 17l-72-72a12 12 0 0 1 0-17l72-72a12 12 0 0 1 17 17L69 116h147a12 12 0 0 1 12 12"/></svg>',
-
-				'»' => '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="m144.49 136.49l-80 80a12 12 0 0 1-17-17L119 128L47.51 56.49a12 12 0 0 1 17-17l80 80a12 12 0 0 1-.02 17m80-17l-80-80a12 12 0 1 0-17 17L199 128l-71.52 71.51a12 12 0 0 0 17 17l80-80a12 12 0 0 0 .01-17Z"/></svg>',
-
-				'«' => '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M208.49 199.51a12 12 0 0 1-17 17l-80-80a12 12 0 0 1 0-17l80-80a12 12 0 0 1 17 17L137 128ZM57 128l71.52-71.51a12 12 0 0 0-17-17l-80 80a12 12 0 0 0 0 17l80 80a12 12 0 0 0 17-17Z"/></svg>',
-			);
-
-
-		// Output
-
-			return str_replace(
-				array(
-					'→', '&rarr;',
-					'←', '&larr;',
-					'»', '&raquo;',
-					'«', '&laquo;',
-				),
-				array(
-					$svg['→'], $svg['→'],
-					$svg['←'], $svg['←'],
-					$svg['»'], $svg['»'],
-					$svg['«'], $svg['«'],
-				),
-				$html
-			);
-
-	} // /replace_arrows
-
-	/**
 	 * Block output modification: Pagination block.
 	 *
 	 * Adding descriptive ARIA labels and helpful current page info data attributes.
@@ -112,7 +70,8 @@ class Pagination implements Component_Interface {
 	 * We need to modify output HTML as there is no filter for `paginate_links()` args.
 	 * @link  https://developer.wordpress.org/reference/functions/paginate_links/
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.0.7
 	 *
 	 * @param  string $block_content  The rendered content. Default null.
 	 * @param  array  $block          The block being rendered.
@@ -209,7 +168,7 @@ class Pagination implements Component_Interface {
 					$block_content = $html->get_updated_html();
 
 					if ( stripos( $block_content, 'is-arrow-' ) ) {
-						$block_content = self::replace_arrows( $block_content );
+						$block_content = Arrow::replace( $block_content );
 					}
 					break;
 			}