+++ parts/admin/welcome-demo.php
@@ -7,7 +7,8 @@
  * @package    Zooey
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.7
  */
 
 // Exit if accessed directly.
@@ -50,5 +51,26 @@ if ( ! class_exists( 'WebManDesign\Zooey\Welcome\Component' ) ) {
 	<hr>
 
 	<p><a href="https://webmandesign.github.io/docs/zooey/#demo-content"><?php esc_html_e( 'Use page block patterns &rarr;', 'zooey' ); ?></a></p>
+
+	<?php if ( get_option( 'fresh_site' ) ) : ?>
+	<div class="welcome__section--starter-content">
+
+		<h3>
+			<span class="welcome__icon dashicons dashicons-edit-page"></span>
+			<?php esc_html_e( 'Starter content', 'zooey' ); ?>
+		</h3>
+
+		<p>
+			<?php esc_html_e( 'The theme also provides a simple starter content.', 'zooey' ); ?>
+			<?php esc_html_e( 'If you haven\'t created any content yet, open the customizer to preview the starter content.', 'zooey' ); ?>
+			<?php esc_html_e( 'Once you publish your customizer settings, the starter content will be imported into your website and you can change it to your needs.', 'zooey' ); ?>
+		</p>
+
+		<p><em><small><?php esc_html_e( '(WordPress starter content works with fresh installation of WordPress only.)', 'zooey' ); ?></small></em>		</p>
+
+		<p><a class="button button-hero" href="<?php echo esc_url( admin_url( 'customize.php' ) ); ?>"><?php esc_html_e( 'Preview starter content', 'zooey' ); ?></a></p>
+
+	</div>
+	<?php endif; ?>
 
 </div>